easy_box_backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
alembic
psycopg2-binary
python-dotenv
python-multipart
gunicorn
sqladmin
itsdangerous
requests
cloudinary
Babel
speaklater